+++ spec/active_job/uniqueness/strategies/until_and_while_executing_spec.rb
@@ -194,4 +194,96 @@ describe ':until_and_while_executing strategy', type: :integration do
       end
     end
   end
+
+  describe 'lock keys' do
+    let(:job) { job_class.new(2, 1) }
+
+    describe 'on enqueuing' do
+      before { job.lock_strategy.before_enqueue }
+
+      context 'when the job has no custom #lock_key defined' do
+        let(:job_class) do
+          stub_active_job_class do
+            unique :until_and_while_executing
+
+            def perform(number1, number2)
+              number1 / number2
+            end
+          end
+        end
+
+        it 'locks the job with the default lock key' do
+          expect(locks.size).to eq 1
+          expect(locks.first).to match(/\Aactivejob_uniqueness:my_job:[^:]+\z/)
+        end
+      end
+
+      context 'when the job has a custom #lock_key defined' do
+        let(:job_class) do
+          stub_active_job_class do
+            unique :until_and_while_executing
+
+            def perform(number1, number2)
+              number1 / number2
+            end
+
+            def lock_key
+              'activejob_uniqueness:whatever'
+            end
+          end
+        end
+
+        it 'locks the job with the custom runtime lock key' do
+          expect(locks.size).to eq 1
+          expect(locks.first).to eq 'activejob_uniqueness:whatever'
+        end
+      end
+    end
+
+    describe 'while executing' do
+      before { job.lock_strategy.before_perform }
+
+      context 'when the job has no custom #runtime_lock_key defined' do
+        let(:job_class) do
+          stub_active_job_class do
+            unique :until_and_while_executing
+
+            def perform(number1, number2)
+              number1 / number2
+            end
+          end
+        end
+
+        it 'locks the job with the default runtime lock key' do
+          job.lock_strategy.around_perform lambda {
+            expect(locks.size).to eq 1
+            expect(locks.first).to match(/\Aactivejob_uniqueness:my_job:[^:]+:runtime\z/)
+          }
+        end
+      end
+
+      context 'when the job has a custom #runtime_lock_key defined' do
+        let(:job_class) do
+          stub_active_job_class do
+            unique :until_and_while_executing
+
+            def perform(number1, number2)
+              number1 / number2
+            end
+
+            def runtime_lock_key
+              'activejob_uniqueness:whatever'
+            end
+          end
+        end
+
+        it 'locks the job with the custom runtime lock key' do
+          job.lock_strategy.around_perform lambda {
+            expect(locks.size).to eq 1
+            expect(locks.first).to eq 'activejob_uniqueness:whatever'
+          }
+        end
+      end
+    end
+  end
 end